Added authentication and authorization to the API

+++ routes/avengers.js
@@ -1,6 +1,9 @@
 const express = require("express");
+const jwt = require("jsonwebtoken");
 const Avenger = require("../models/avenger");
 const router = express.Router();
+
+SECRET_KEY = "123456789";
 
 // GET ALL
 router.get("/", async (req, res) => {
@@ -21,6 +24,11 @@ router.get("/", async (req, res) => {
 // ==      comparing the value "2" == 2
 // ===     comparing the value and the data type "2" === 2
 router.get("/:avengerId", async (req, res) => {
+  try {
+    jwt.verify(token, SECRET_KEY);
+  } catch (e) {
+    return res.status(400).send("Invalid token");
+  }
   let avenger = await Avenger.findById(req.params.avengerId);
   if (!avenger) {
     return res.status(404).send("The given ID does not exist on our system");
@@ -29,6 +37,14 @@ router.get("/:avengerId", async (req, res) => {
 });
 
 router.post("/", async (req, res) => {
+  const token = req.header("x-jwt-token");
+  if (!token) return res.status(401).send("Access is denied. No token found");
+
+  try {
+    jwt.verify(token, SECRET_KEY);
+  } catch (e) {
+    return res.status(400).send("Invalid token");
+  }
   // if (!req.body.avengerName) {
   //   return res.status(400).send("Not all mandatory values are sent");
   // }
@@ -66,7 +82,22 @@ router.put("/:avengerId", async (req, res) => {
 });
 
 router.delete("/:avengerId", async (req, res) => {
-  let avenger = await Avenger.findOneAndDelete({ _id: req.params.avengerId})
+  const token = req.header("x-jwt-token");
+  if (!token) return res.status(401).send("Access is denied. No token found");
+
+  // Checking whether valid and authenticated
+  try {
+    jwt.verify(token, SECRET_KEY);
+  } catch (e) {
+    return res.status(400).send("Invalid token");
+  }
+
+  let decoded = jwt.decode(token, SECRET_KEY);
+  if (!decoded.isAdmin) {
+    return res.status(403).send("Forbidden. You don't have access to this endpoint")
+  }
+
+  let avenger = await Avenger.findOneAndDelete({ _id: req.params.avengerId });
   if (!avenger) {
     return res.status(404).send("The given ID does not exist on our system");
   }
